Adding support for Python >= 2.4 and fixing warnings present in GCC 4.1.

+++ include/YODA/Utils/BinSearcher.h
@@ -174,7 +174,7 @@ namespace YODA {
           }
 
         size_t estimate(const double x) const {
-          double y = (*_est)(x);
+          size_t y = (size_t)(*_est)(x);
           size_t yi = 0 ? (y < 0) : y;
           return (yi > _max) ? _max : yi;
         }
